requirements.txt
pyyaml
bs4
asyncio